75% of the interface is ready. Reorganization of the button system.
The button system has been completely reorganized - previously it was the layout that buttons were added to. Further, referring to it, the buttons were updated. Button clicks were caught not from the layout.

Now, instead of the layout, a multi-page interface has been added. This allows you to put pointers to buttons in a separate array, and work through it. The array of buttons is located on the widget, which is declared and initialized in the code. The widget is added to QStackedWidget, which allows you to split the interface into pages and work with them separately.

The only problem is the impossibility to create elements on an already finished page. One element - yes, but to add many - you need a widget. Actually, adding elements to the widget, we add elements to the page with the game itself.

Полностью реорганизована система кнопок - ранее это был макет, на который добавлялись кнопки. Далее, обращаясь к нему, производилось обновление кнопок. Отлов нажатий на кнопки производился не от макета.

Сейчас вместо макета добавлен многостраничный интерфейс. Это позволяет вынести указатели на кнопки в отдельный массив, и работать через него. Массив кнопок распологается на виджете, который объявлен и инициализорван в коде. Виджет добавлен на QStackedWidget, что позволяет разделить интерфейс на  страницы, и работать с ними по отдельности.

Единственная проблема - это невозможность создавать элементы на уже готовой странице. Один элемент - да, но чтобы добавить много - нужен виджет. Собственно, добавляя элементы на виджет, мы добавляем элементы на страницу с самой игрой.